[AIP-21] Move sensors.hdfs_sensor providers.apache.hdfs.sensors.hdfs
