+++ tests/providers/apache/hdfs/sensors/test_hdfs.py
@@ -22,8 +22,71 @@ import unittest
 from datetime import timedelta
 
 from airflow.exceptions import AirflowSensorTimeout
-from airflow.providers.apache.hdfs.sensors.hdfs import HdfsSensorFolder, HdfsSensorRegex
+from airflow.providers.apache.hdfs.sensors.hdfs import HdfsSensor, HdfsSensorFolder, HdfsSensorRegex
+from airflow.utils.timezone import datetime
 from tests.test_utils.hdfs_utils import FakeHDFSHook
+
+DEFAULT_DATE = datetime(2015, 1, 1)
+TEST_DAG_ID = 'unit_test_dag'
+
+
+class TestHdfsSensor(unittest.TestCase):
+
+    def setUp(self):
+        self.hook = FakeHDFSHook
+
+    def test_legacy_file_exist(self):
+        """
+        Test the legacy behaviour
+        :return:
+        """
+        # When
+        task = HdfsSensor(task_id='Should_be_file_legacy',
+                          filepath='/datadirectory/datafile',
+                          timeout=1,
+                          retry_delay=timedelta(seconds=1),
+                          poke_interval=1,
+                          hook=self.hook)
+        task.execute(None)
+
+        # Then
+        # Nothing happens, nothing is raised exec is ok
+
+    def test_legacy_file_exist_but_filesize(self):
+        """
+        Test the legacy behaviour with the filesize
+        :return:
+        """
+        # When
+        task = HdfsSensor(task_id='Should_be_file_legacy',
+                          filepath='/datadirectory/datafile',
+                          timeout=1,
+                          file_size=20,
+                          retry_delay=timedelta(seconds=1),
+                          poke_interval=1,
+                          hook=self.hook)
+
+        # When
+        # Then
+        with self.assertRaises(AirflowSensorTimeout):
+            task.execute(None)
+
+    def test_legacy_file_does_not_exists(self):
+        """
+        Test the legacy behaviour
+        :return:
+        """
+        task = HdfsSensor(task_id='Should_not_be_file_legacy',
+                          filepath='/datadirectory/not_existing_file_or_directory',
+                          timeout=1,
+                          retry_delay=timedelta(seconds=1),
+                          poke_interval=1,
+                          hook=self.hook)
+
+        # When
+        # Then
+        with self.assertRaises(AirflowSensorTimeout):
+            task.execute(None)
 
 
 class TestHdfsSensorFolder(unittest.TestCase):
